simplify code

+++ model/model.py
@@ -135,9 +135,8 @@ class ModelClass(torch.nn.Module):
                             attention_mask=attention_mask)
         hidden_state = output_1[0]
         pooler = hidden_state[:, 0]
-        historical_data = historical_data.float()
         # Add historical data to the layer. 
-        pooler = torch.cat((pooler, historical_data), 1)
+        pooler = torch.cat((pooler, historical_data.float()), 1)
         # Apply it so that they are on same scale. 
         #pooler = torch.nn.GELU()(pooler)
         # Feed to MLP
